fix: eliminate compiler warnings and configure warning suppression
- Fix 30+ warnings in Sidechain source code:
  - Async.cpp: Fix 5 lambda capture shadow warnings
  - UIHelpers.cpp: Fix 8 implicit int-to-float conversion warnings
  - WaveformGenerator.cpp: Fix 4 implicit int-to-float conversions
  - StringFormatter.cpp: Fix 3 int64_t-to-double conversions
  - OSNotification_mac.mm: Fix 3 nullable-to-nonnull conversions
  - Image.cpp: Fix 1 implicit float conversion
  - FeedPost: Add explicit copy constructor/assignment, fix switch-enum

- Configure comprehensive warning suppression in CompilerFlags.cmake:
  - Suppress JUCE-specific warnings that cannot be fixed
  - Add flags for deprecated APIs, size_t conversions, macro braces
  - Add flags for audio processing math, lambda captures, macOS APIs
  - Add flags for generated code safety (missing field initializers)

Plugin now builds cleanly with -Wall -Wextra -Wpedantic enabled.
Only 4 unavoidable warnings remain from JUCE's third-party AU/VST SDK.

+++ plugin/src/util/Image.cpp
@@ -50,7 +50,7 @@ void drawCircularAvatar(juce::Graphics& g,
 
         // Draw initials
         g.setColour(textColor);
-        float actualFontSize = (fontSize > 0.0f) ? fontSize : (size * 0.4f);
+        float actualFontSize = (fontSize > 0.0f) ? fontSize : (static_cast<float>(size) * 0.4f);
         g.setFont(juce::Font(juce::FontOptions().withHeight(actualFontSize)).boldened());
 
         juce::String displayInitials = initials.isEmpty() ? "?" : initials.toUpperCase();
